refactor: コンポーネントとミックスインを別のディレクトリで管理するようにする
