+++ app/frontend/packs/javascript/components/step_circle_component.js
