+++ app/frontend/packs/javascript/components/update_task_form_component.js
@@ -1,4 +1,4 @@
-import taskFormRequestable from './task_form_requestable';
+import taskFormRequestable from '../modules/task_form_requestable';
 
 export default {
   props: ['task'],
