+++ app/frontend/packs/javascript/steps_vue.js
@@ -1,7 +1,7 @@
 import Vue from 'vue/dist/vue.esm.js';
 import '../stylesheets/application.scss';
 import {MdButton} from 'vue-material/dist/components';
-import stepCircleComponent from './modules/step_circle_component';
+import stepCircleComponent from './components/step_circle_component';
 
 Vue.use(MdButton);
